Fixed project structure (samples, test, src)
