--- src/AzureFunctionExtensions.Test/RedisDatabaseMock.cs
@@ -1,1436 +1,0 @@
-﻿using StackExchange.Redis;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Fbeltrao.AzureFunctionExtensions.Test
-{
-    public class RedisDatabaseMock : IDatabase
-    {
-
-        Dictionary<string, RedisValue> keyCollection = new Dictionary<string, RedisValue>();
-        Dictionary<string, List<RedisValue>> listCollection = new Dictionary<string, List<RedisValue>>();
-
-        public int Database => throw new NotImplementedException();
-
-        public ConnectionMultiplexer Multiplexer => throw new NotImplementedException();
-
-        public IBatch CreateBatch(object asyncState = null)
-        {
-            throw new NotImplementedException();
-        }
-
-        public ITransaction CreateTransaction(object asyncState = null)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue DebugObject(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> DebugObjectAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long HashDecrement(RedisKey key, RedisValue hashField, long value = 1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public double HashDecrement(RedisKey key, RedisValue hashField, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> HashDecrementAsync(RedisKey key, RedisValue hashField, long value = 1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<double> HashDecrementAsync(RedisKey key, RedisValue hashField, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool HashDelete(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long HashDelete(RedisKey key, RedisValue[] hashFields, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> HashDeleteAsync(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> HashDeleteAsync(RedisKey key, RedisValue[] hashFields, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool HashExists(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> HashExistsAsync(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue HashGet(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] HashGet(RedisKey key, RedisValue[] hashFields, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public HashEntry[] HashGetAll(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<HashEntry[]> HashGetAllAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> HashGetAsync(RedisKey key, RedisValue hashField, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> HashGetAsync(RedisKey key, RedisValue[] hashFields, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long HashIncrement(RedisKey key, RedisValue hashField, long value = 1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public double HashIncrement(RedisKey key, RedisValue hashField, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> HashIncrementAsync(RedisKey key, RedisValue hashField, long value = 1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<double> HashIncrementAsync(RedisKey key, RedisValue hashField, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] HashKeys(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> HashKeysAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long HashLength(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> HashLengthAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<HashEntry> HashScan(RedisKey key, RedisValue pattern, int pageSize, CommandFlags flags)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<HashEntry> HashScan(RedisKey key, RedisValue pattern = default(RedisValue), int pageSize = 10, long cursor = 0, int pageOffset = 0, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void HashSet(RedisKey key, HashEntry[] hashFields, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool HashSet(RedisKey key, RedisValue hashField, RedisValue value, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task HashSetAsync(RedisKey key, HashEntry[] hashFields, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> HashSetAsync(RedisKey key, RedisValue hashField, RedisValue value, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] HashValues(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> HashValuesAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool HyperLogLogAdd(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool HyperLogLogAdd(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> HyperLogLogAddAsync(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> HyperLogLogAddAsync(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long HyperLogLogLength(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long HyperLogLogLength(RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> HyperLogLogLengthAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> HyperLogLogLengthAsync(RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void HyperLogLogMerge(RedisKey destination, RedisKey first, RedisKey second, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void HyperLogLogMerge(RedisKey destination, RedisKey[] sourceKeys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task HyperLogLogMergeAsync(RedisKey destination, RedisKey first, RedisKey second, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task HyperLogLogMergeAsync(RedisKey destination, RedisKey[] sourceKeys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public EndPoint IdentifyEndpoint(RedisKey key = default(RedisKey), CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<EndPoint> IdentifyEndpointAsync(RedisKey key = default(RedisKey), CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool IsConnected(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool KeyDelete(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long KeyDelete(RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> KeyDeleteAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> KeyDeleteAsync(RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public byte[] KeyDump(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<byte[]> KeyDumpAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool KeyExists(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> KeyExistsAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool KeyExpire(RedisKey key, TimeSpan? expiry, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool KeyExpire(RedisKey key, DateTime? expiry, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> KeyExpireAsync(RedisKey key, TimeSpan? expiry, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> KeyExpireAsync(RedisKey key, DateTime? expiry, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void KeyMigrate(RedisKey key, EndPoint toServer, int toDatabase = 0, int timeoutMilliseconds = 0, MigrateOptions migrateOptions = MigrateOptions.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task KeyMigrateAsync(RedisKey key, EndPoint toServer, int toDatabase = 0, int timeoutMilliseconds = 0, MigrateOptions migrateOptions = MigrateOptions.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool KeyMove(RedisKey key, int database, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> KeyMoveAsync(RedisKey key, int database, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool KeyPersist(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> KeyPersistAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisKey KeyRandom(CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisKey> KeyRandomAsync(CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool KeyRename(RedisKey key, RedisKey newKey, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> KeyRenameAsync(RedisKey key, RedisKey newKey, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void KeyRestore(RedisKey key, byte[] value, TimeSpan? expiry = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task KeyRestoreAsync(RedisKey key, byte[] value, TimeSpan? expiry = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public TimeSpan? KeyTimeToLive(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<TimeSpan?> KeyTimeToLiveAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisType KeyType(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisType> KeyTypeAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue ListGetByIndex(RedisKey key, long index, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> ListGetByIndexAsync(RedisKey key, long index, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long ListInsertAfter(RedisKey key, RedisValue pivot, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> ListInsertAfterAsync(RedisKey key, RedisValue pivot, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long ListInsertBefore(RedisKey key, RedisValue pivot, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> ListInsertBeforeAsync(RedisKey key, RedisValue pivot, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue ListLeftPop(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> ListLeftPopAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long ListLeftPush(RedisKey key, RedisValue value, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            if (!listCollection.TryGetValue(key, out var list))
-            {
-                list = new List<RedisValue>();
-                listCollection[key] = list;
-            }
-
-            list.Insert(0, value);
-
-            return list.Count;
-        }
-
-        public long ListLeftPush(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            if (!listCollection.TryGetValue(key, out var list))
-            {
-                list = new List<RedisValue>();
-                listCollection[key] = list;
-            }
-
-            list.InsertRange(0, values);
-
-            return list.Count;
-        }
-
-        public Task<long> ListLeftPushAsync(RedisKey key, RedisValue value, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            if (!listCollection.TryGetValue(key, out var list))
-            {
-                list = new List<RedisValue>();
-                listCollection[key] = list;
-            }
-
-            list.Insert(0, value);
-
-            return Task.FromResult((long)list.Count);
-        }
-
-        public Task<long> ListLeftPushAsync(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            if (!listCollection.TryGetValue(key, out var list))
-            {
-                list = new List<RedisValue>();
-                listCollection[key] = list;
-            }
-
-            list.InsertRange(0, values);
-
-            return Task.FromResult((long)list.Count);
-        }
-
-        public long ListLength(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> ListLengthAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] ListRange(RedisKey key, long start = 0, long stop = -1, CommandFlags flags = CommandFlags.None)
-        {
-            RedisValue[] result = null;
-            if (this.listCollection.TryGetValue(key, out var list))
-            {
-                if (start < list.Count)
-                {
-                    if (stop <= 0 && stop < list.Count)
-                    {
-                        var count = stop < 0 ? list.Count - start : stop - start;
-                        result = list.GetRange((int)start, (int)count).ToArray();
-                    }
-                }
-            }
-
-            return (result ?? new RedisValue[0]);
-        }
-
-        public Task<RedisValue[]> ListRangeAsync(RedisKey key, long start = 0, long stop = -1, CommandFlags flags = CommandFlags.None)
-        {
-            RedisValue[] result = null;
-            if (this.listCollection.TryGetValue(key, out var list))
-            {
-                if (start < list.Count)
-                {
-                    if (stop <= 0 && stop < list.Count)
-                    {
-                        var count = stop < 0 ? list.Count - start : stop - start;
-                        result = list.GetRange((int)start, (int)count).ToArray();
-                    }
-                }
-            }
-
-            return Task.FromResult(result ?? new RedisValue[0]);
-        }
-
-        public long ListRemove(RedisKey key, RedisValue value, long count = 0, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> ListRemoveAsync(RedisKey key, RedisValue value, long count = 0, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue ListRightPop(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> ListRightPopAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue ListRightPopLeftPush(RedisKey source, RedisKey destination, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> ListRightPopLeftPushAsync(RedisKey source, RedisKey destination, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long ListRightPush(RedisKey key, RedisValue value, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            if (!listCollection.TryGetValue(key, out var list))
-            {
-                list = new List<RedisValue>();
-                listCollection[key] = list;
-            }
-
-            list.Add(value);
-
-            return list.Count;
-        }
-
-        public long ListRightPush(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            if (!listCollection.TryGetValue(key, out var list))
-            {
-                list = new List<RedisValue>();
-                listCollection[key] = list;
-            }
-
-            list.AddRange(values);
-
-            return list.Count;
-        }     
-
-
-        public Task<long> ListRightPushAsync(RedisKey key, RedisValue value, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            if (!listCollection.TryGetValue(key, out var list))
-            {
-                list = new List<RedisValue>();
-                listCollection[key] = list;
-            }
-
-            list.Add(value);
-
-            return Task.FromResult((long)list.Count);
-        }
-
-        public Task<long> ListRightPushAsync(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            if (!listCollection.TryGetValue(key, out var list))
-            {
-                list = new List<RedisValue>();
-                listCollection[key] = list;
-            }
-
-            list.AddRange(values);
-
-            return Task.FromResult((long)list.Count);
-        }
-
-        public void ListSetByIndex(RedisKey key, long index, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task ListSetByIndexAsync(RedisKey key, long index, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void ListTrim(RedisKey key, long start, long stop, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task ListTrimAsync(RedisKey key, long start, long stop, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool LockExtend(RedisKey key, RedisValue value, TimeSpan expiry, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> LockExtendAsync(RedisKey key, RedisValue value, TimeSpan expiry, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue LockQuery(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> LockQueryAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool LockRelease(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> LockReleaseAsync(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool LockTake(RedisKey key, RedisValue value, TimeSpan expiry, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> LockTakeAsync(RedisKey key, RedisValue value, TimeSpan expiry, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public TimeSpan Ping(CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<TimeSpan> PingAsync(CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long Publish(RedisChannel channel, RedisValue message, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> PublishAsync(RedisChannel channel, RedisValue message, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisResult ScriptEvaluate(string script, RedisKey[] keys = null, RedisValue[] values = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisResult ScriptEvaluate(byte[] hash, RedisKey[] keys = null, RedisValue[] values = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisResult ScriptEvaluate(LuaScript script, object parameters = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisResult ScriptEvaluate(LoadedLuaScript script, object parameters = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisResult> ScriptEvaluateAsync(string script, RedisKey[] keys = null, RedisValue[] values = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisResult> ScriptEvaluateAsync(byte[] hash, RedisKey[] keys = null, RedisValue[] values = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisResult> ScriptEvaluateAsync(LuaScript script, object parameters = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisResult> ScriptEvaluateAsync(LoadedLuaScript script, object parameters = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool SetAdd(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SetAdd(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> SetAddAsync(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SetAddAsync(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] SetCombine(SetOperation operation, RedisKey first, RedisKey second, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] SetCombine(SetOperation operation, RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SetCombineAndStore(SetOperation operation, RedisKey destination, RedisKey first, RedisKey second, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SetCombineAndStore(SetOperation operation, RedisKey destination, RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SetCombineAndStoreAsync(SetOperation operation, RedisKey destination, RedisKey first, RedisKey second, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SetCombineAndStoreAsync(SetOperation operation, RedisKey destination, RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> SetCombineAsync(SetOperation operation, RedisKey first, RedisKey second, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> SetCombineAsync(SetOperation operation, RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool SetContains(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> SetContainsAsync(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SetLength(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SetLengthAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] SetMembers(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> SetMembersAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool SetMove(RedisKey source, RedisKey destination, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> SetMoveAsync(RedisKey source, RedisKey destination, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue SetPop(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> SetPopAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue SetRandomMember(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> SetRandomMemberAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] SetRandomMembers(RedisKey key, long count, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> SetRandomMembersAsync(RedisKey key, long count, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool SetRemove(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SetRemove(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> SetRemoveAsync(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SetRemoveAsync(RedisKey key, RedisValue[] values, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<RedisValue> SetScan(RedisKey key, RedisValue pattern, int pageSize, CommandFlags flags)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<RedisValue> SetScan(RedisKey key, RedisValue pattern = default(RedisValue), int pageSize = 10, long cursor = 0, int pageOffset = 0, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] Sort(RedisKey key, long skip = 0, long take = -1, Order order = Order.Ascending, SortType sortType = SortType.Numeric, RedisValue by = default(RedisValue), RedisValue[] get = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortAndStore(RedisKey destination, RedisKey key, long skip = 0, long take = -1, Order order = Order.Ascending, SortType sortType = SortType.Numeric, RedisValue by = default(RedisValue), RedisValue[] get = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortAndStoreAsync(RedisKey destination, RedisKey key, long skip = 0, long take = -1, Order order = Order.Ascending, SortType sortType = SortType.Numeric, RedisValue by = default(RedisValue), RedisValue[] get = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> SortAsync(RedisKey key, long skip = 0, long take = -1, Order order = Order.Ascending, SortType sortType = SortType.Numeric, RedisValue by = default(RedisValue), RedisValue[] get = null, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool SortedSetAdd(RedisKey key, RedisValue member, double score, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortedSetAdd(RedisKey key, SortedSetEntry[] values, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> SortedSetAddAsync(RedisKey key, RedisValue member, double score, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortedSetAddAsync(RedisKey key, SortedSetEntry[] values, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortedSetCombineAndStore(SetOperation operation, RedisKey destination, RedisKey first, RedisKey second, Aggregate aggregate = Aggregate.Sum, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortedSetCombineAndStore(SetOperation operation, RedisKey destination, RedisKey[] keys, double[] weights = null, Aggregate aggregate = Aggregate.Sum, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortedSetCombineAndStoreAsync(SetOperation operation, RedisKey destination, RedisKey first, RedisKey second, Aggregate aggregate = Aggregate.Sum, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortedSetCombineAndStoreAsync(SetOperation operation, RedisKey destination, RedisKey[] keys, double[] weights = null, Aggregate aggregate = Aggregate.Sum, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public double SortedSetDecrement(RedisKey key, RedisValue member, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<double> SortedSetDecrementAsync(RedisKey key, RedisValue member, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public double SortedSetIncrement(RedisKey key, RedisValue member, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<double> SortedSetIncrementAsync(RedisKey key, RedisValue member, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortedSetLength(RedisKey key, double min = double.NegativeInfinity, double max = double.PositiveInfinity, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortedSetLengthAsync(RedisKey key, double min = double.NegativeInfinity, double max = double.PositiveInfinity, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortedSetLengthByValue(RedisKey key, RedisValue min, RedisValue max, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortedSetLengthByValueAsync(RedisKey key, RedisValue min, RedisValue max, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] SortedSetRangeByRank(RedisKey key, long start = 0, long stop = -1, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> SortedSetRangeByRankAsync(RedisKey key, long start = 0, long stop = -1, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public SortedSetEntry[] SortedSetRangeByRankWithScores(RedisKey key, long start = 0, long stop = -1, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<SortedSetEntry[]> SortedSetRangeByRankWithScoresAsync(RedisKey key, long start = 0, long stop = -1, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] SortedSetRangeByScore(RedisKey key, double start = double.NegativeInfinity, double stop = double.PositiveInfinity, Exclude exclude = Exclude.None, Order order = Order.Ascending, long skip = 0, long take = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> SortedSetRangeByScoreAsync(RedisKey key, double start = double.NegativeInfinity, double stop = double.PositiveInfinity, Exclude exclude = Exclude.None, Order order = Order.Ascending, long skip = 0, long take = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public SortedSetEntry[] SortedSetRangeByScoreWithScores(RedisKey key, double start = double.NegativeInfinity, double stop = double.PositiveInfinity, Exclude exclude = Exclude.None, Order order = Order.Ascending, long skip = 0, long take = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<SortedSetEntry[]> SortedSetRangeByScoreWithScoresAsync(RedisKey key, double start = double.NegativeInfinity, double stop = double.PositiveInfinity, Exclude exclude = Exclude.None, Order order = Order.Ascending, long skip = 0, long take = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue[] SortedSetRangeByValue(RedisKey key, RedisValue min = default(RedisValue), RedisValue max = default(RedisValue), Exclude exclude = Exclude.None, long skip = 0, long take = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue[]> SortedSetRangeByValueAsync(RedisKey key, RedisValue min = default(RedisValue), RedisValue max = default(RedisValue), Exclude exclude = Exclude.None, long skip = 0, long take = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long? SortedSetRank(RedisKey key, RedisValue member, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long?> SortedSetRankAsync(RedisKey key, RedisValue member, Order order = Order.Ascending, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool SortedSetRemove(RedisKey key, RedisValue member, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortedSetRemove(RedisKey key, RedisValue[] members, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> SortedSetRemoveAsync(RedisKey key, RedisValue member, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortedSetRemoveAsync(RedisKey key, RedisValue[] members, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortedSetRemoveRangeByRank(RedisKey key, long start, long stop, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortedSetRemoveRangeByRankAsync(RedisKey key, long start, long stop, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortedSetRemoveRangeByScore(RedisKey key, double start, double stop, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortedSetRemoveRangeByScoreAsync(RedisKey key, double start, double stop, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long SortedSetRemoveRangeByValue(RedisKey key, RedisValue min, RedisValue max, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> SortedSetRemoveRangeByValueAsync(RedisKey key, RedisValue min, RedisValue max, Exclude exclude = Exclude.None, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<SortedSetEntry> SortedSetScan(RedisKey key, RedisValue pattern, int pageSize, CommandFlags flags)
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<SortedSetEntry> SortedSetScan(RedisKey key, RedisValue pattern = default(RedisValue), int pageSize = 10, long cursor = 0, int pageOffset = 0, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public double? SortedSetScore(RedisKey key, RedisValue member, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<double?> SortedSetScoreAsync(RedisKey key, RedisValue member, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long StringAppend(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> StringAppendAsync(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long StringBitCount(RedisKey key, long start = 0, long end = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> StringBitCountAsync(RedisKey key, long start = 0, long end = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long StringBitOperation(Bitwise operation, RedisKey destination, RedisKey first, RedisKey second = default(RedisKey), CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long StringBitOperation(Bitwise operation, RedisKey destination, RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> StringBitOperationAsync(Bitwise operation, RedisKey destination, RedisKey first, RedisKey second = default(RedisKey), CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> StringBitOperationAsync(Bitwise operation, RedisKey destination, RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long StringBitPosition(RedisKey key, bool bit, long start = 0, long end = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> StringBitPositionAsync(RedisKey key, bool bit, long start = 0, long end = -1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long StringDecrement(RedisKey key, long value = 1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public double StringDecrement(RedisKey key, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> StringDecrementAsync(RedisKey key, long value = 1, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<double> StringDecrementAsync(RedisKey key, double value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue StringGet(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            if (keyCollection.TryGetValue(key, out var value))
-                return value;
-
-            return RedisValue.Null;
-        }
-
-        public RedisValue[] StringGet(RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            var result = new List<RedisValue>();
-            foreach (var key in keys)
-            {
-                if (keyCollection.TryGetValue(key, out var value))
-                    result.Add(value);
-                else
-                    result.Add(RedisValue.Null);
-            }
-
-            return result.ToArray();
-        }
-
-        public Task<RedisValue> StringGetAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            if (keyCollection.TryGetValue(key, out var value))
-                return Task.FromResult(value);
-
-            return Task.FromResult(RedisValue.Null);
-        }
-
-        public Task<RedisValue[]> StringGetAsync(RedisKey[] keys, CommandFlags flags = CommandFlags.None)
-        {
-            var result = new List<RedisValue>();
-            foreach (var key in keys)
-            {
-                if (keyCollection.TryGetValue(key, out var value))
-                    result.Add(value);
-                else
-                    result.Add(RedisValue.Null);
-            }
-
-            return Task.FromResult(result.ToArray());
-        }
-
-        public bool StringGetBit(RedisKey key, long offset, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> StringGetBitAsync(RedisKey key, long offset, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue StringGetRange(RedisKey key, long start, long end, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> StringGetRangeAsync(RedisKey key, long start, long end, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        
-        public RedisValue StringGetSet(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            keyCollection[key.ToString()] = value;
-            return value;
-        }
-
-        public Task<RedisValue> StringGetSetAsync(RedisKey key, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValueWithExpiry StringGetWithExpiry(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValueWithExpiry> StringGetWithExpiryAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long StringIncrement(RedisKey key, long value = 1, CommandFlags flags = CommandFlags.None)
-        {
-            var existingValue = 0L;
-            if (keyCollection.TryGetValue(key, out var existing))
-                existingValue = (long)existing;
-
-            var newValue = value + existingValue;
-
-            keyCollection[key] = newValue;
-            return newValue;
-        }
-
-        public double StringIncrement(RedisKey key, double value, CommandFlags flags = CommandFlags.None)
-        {
-            var existingValue = 0d;
-            if (keyCollection.TryGetValue(key, out var existing))
-                existingValue = (double)existing;
-
-            var newValue = value + existingValue;
-
-            keyCollection[key] = newValue;
-            return newValue;
-        }
-
-        public Task<long> StringIncrementAsync(RedisKey key, long value = 1, CommandFlags flags = CommandFlags.None)
-        {
-            var existingValue = 0L;
-            if (keyCollection.TryGetValue(key, out var existing))
-                existingValue = (long)existing;
-
-            var newValue = value + existingValue;
-
-            keyCollection[key] = newValue;
-
-            return Task.FromResult(newValue);
-        }
-
-        public Task<double> StringIncrementAsync(RedisKey key, double value, CommandFlags flags = CommandFlags.None)
-        {
-            var existingValue = 0d;
-            if (keyCollection.TryGetValue(key, out var existing))
-                existingValue = (double)existing;
-
-            var newValue = value + existingValue;
-
-            keyCollection[key] = newValue;
-
-            return Task.FromResult(newValue);
-        }
-
-        public long StringLength(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<long> StringLengthAsync(RedisKey key, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool StringSet(RedisKey key, RedisValue value, TimeSpan? expiry = null, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            this.keyCollection[key] = value;
-            return true;
-        }
-
-        public bool StringSet(KeyValuePair<RedisKey, RedisValue>[] values, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            foreach (var de in values)
-            {
-                this.keyCollection[de.Key] = de.Value;
-            }
-
-            return true;
-        }
-
-        public Task<bool> StringSetAsync(RedisKey key, RedisValue value, TimeSpan? expiry = null, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            this.keyCollection[key] = value;
-
-            return Task.FromResult<bool>(true);
-        }
-
-        public Task<bool> StringSetAsync(KeyValuePair<RedisKey, RedisValue>[] values, When when = When.Always, CommandFlags flags = CommandFlags.None)
-        {
-            foreach (var de in values)
-            {
-                this.keyCollection[de.Key] = de.Value;
-            }
-
-            return Task.FromResult<bool>(true);
-        }
-
-        public bool StringSetBit(RedisKey key, long offset, bool bit, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<bool> StringSetBitAsync(RedisKey key, long offset, bool bit, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public RedisValue StringSetRange(RedisKey key, long offset, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<RedisValue> StringSetRangeAsync(RedisKey key, long offset, RedisValue value, CommandFlags flags = CommandFlags.None)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool TryWait(Task task)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void Wait(Task task)
-        {
-            throw new NotImplementedException();
-        }
-
-        public T Wait<T>(Task<T> task)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void WaitAll(params Task[] tasks)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
