+++ src/AzureFunctionExtensions/AzureFunctionExtensions.csproj
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <RootNamespace>Fbeltrao.AzureFunctionExtensions</RootNamespace>
+    <AssemblyName>Fbeltrao.AzureFunctionExtensions</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
