--- src/samples/RedisSample/host.json
@@ -1,2 +1,0 @@
-{
-}
