+++ test/AzureFunctionExtensions.Test/EventGridOutputAsyncCollectorTest.cs
